+++ src/components/comics-list/comics-list.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import ComicsList from "./comics-list";
-import comics from "./mocks/comics.json";
+import comics from "../../mocks/comics.json";
 
 const meta: Meta<typeof ComicsList> = {
   component: ComicsList
