+++ src/mocks/comics.json
